backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain and AI dependencies
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# HTTP requests and caching
requests==2.31.0
httpx==0.25.2
cachetools==5.3.2

# Rate limiting
slowapi==0.1.9

# CORS middleware
python-jose[cryptography]==3.3.0
